requirements.txt
tensorflow-cpu
numpy
matplotlib